Create logic for a single round game
This commit completes the coding for a single
round of rock paper scissors. I learned that there
is a very important difference between using
console.log to check a function result rather than
using the variable assigned to the function. One
will recall the function while the other will take
the result from the previous function. I think
this will be an important thing to keep in mind.